# src/main/resources/openapi.yml
openapi: 3.0.0
info:
  title: Recipes API
  version: 1.0.0
  description: API for managing recipes and their ingredients
tags:
  - name: Recipe
  - name: Request
paths:
  /recipe:
    get:
      summary: Get paginated recipes with search
      operationId: searchRecipes
      tags:
        - Recipe
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
            minimum: 0
          description: Zero-based page index
        - name: size
          in: query
          required: true
          schema:
            type: integer
            minimum: 1
          description: The size of the page to be returned
        - name: search
          in: query
          description: Search term to filter recipes
          schema:
            type: string
      responses:
        '200':
          description: Paginated recipes
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedRecipeResponse'
        '400':
          description: Invalid query parameters
        '500':
          description: Internal server error
    post:
      summary: Create a new recipe
      operationId: createRecipe
      tags:
        - Recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeRequest'
      responses:
        '201':
          description: Successfully created recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid request payload
        '500':
          description: Internal server error

  /recipe/{recipeId}:
    get:
      summary: Get recipe by ID
      operationId: getRecipe
      tags:
        - Recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Recipe found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid recipe id supplied
        '404':
          description: Recipe not found
        '500':
          description: Internal server error
    put:
      summary: Update an existing recipe
      operationId: updateRecipe
      tags:
        - Recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RecipeUpdateRequest'
      responses:
        '200':
          description: Successfully updated recipe
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecipeResponse'
        '400':
          description: Invalid request payload
        '404':
          description: Recipe not found
        '500':
          description: Internal server error
    delete:
      summary: Delete recipe by id
      operationId: deleteRecipe
      tags:
        - Recipe
      parameters:
        - name: recipeId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Successfully deleted the recipe
        '400':
          description: Invalid recipe ID supplied
        '404':
          description: Recipe not found
        '500':
          description: Internal server error

  /request/count:
    get:
      summary: Get total count of requests
      operationId: getRequestCount
      tags:
        - Request
      responses:
        '200':
          description: Successfully fetched request count
          content:
            application/json:
              schema:
                type: integer
                format: int64

components:
  schemas:
    RecipeRequest:
      type: object
      required:
        - name
        - description
        - difficulty
        - preparationTime
        - ingredients
      properties:
        name:
          type: string
          maxLength: 255
          example: "Curry"
        description:
          type: string
          example: "Chop onions, fry spices, add vegetables, simmer for 20 minutes..."
        difficulty:
          $ref: '#/components/schemas/Difficulty'
        preparationTime:
          type: integer
          description: Time in minutes
        ingredients:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/IngredientRequest'

    RecipeUpdateRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/RecipeRequest'
        - type: object
          properties:
            ingredients:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/IngredientUpdateRequest'

    RecipeResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/RecipeRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64
            ingredients:
              type: array
              minItems: 1
              items:
                $ref: '#/components/schemas/IngredientResponse'

    IngredientRequest:
      type: object
      required:
        - name
        - quantity
        - unit
      properties:
        name:
          type: string
          maxLength: 60
          example: "Egg"
        quantity:
          type: number
          format: double
        unit:
          $ref: '#/components/schemas/Unit'

    IngredientUpdateRequest:
      type: object
      allOf:
        - $ref: '#/components/schemas/IngredientRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64

    IngredientResponse:
      type: object
      allOf:
        - $ref: '#/components/schemas/IngredientRequest'
        - type: object
          properties:
            id:
              type: integer
              format: int64

    PaginatedRecipeResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/RecipeResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    Difficulty:
      type: string
      enum:
        - EASY
        - MEDIUM
        - HARD

    Unit:
      type: string
      enum:
        - G
        - KG
        - ML
        - CUP
        - TBSP
        - TSP
        - PIECE